requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
psycopg2-binary==2.9.10
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.1 